Change SectionHistories: change card to shader slider
