--- app/_components/SectionHistories/HistoryCard/index.js
@@ -1,3 +1,0 @@
-import HistoryCard from './HistoryCard';
-
-export default HistoryCard;
