+++ app/_components/SectionHistories/HistorySlider/HistorySlider.module.scss
@@ -1,0 +1,213 @@
+@import './app/_styles/variables.scss';
+@import './app/_styles/mixin.scss';
+
+.item {
+  position: relative;
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  min-height: 455px;
+  padding: 30px 30px 18px;
+  box-sizing: border-box;
+  border-radius: 20px;
+  color: var(--third-text-color);
+  text-align: start;
+  overflow: hidden;
+
+  @media (min-width: $desktop-md-width) {
+    min-height: 567px;
+    padding: 6.5% 6.5% 5.6%;
+    border-radius: 50px;
+  }
+  @media (min-width: $desktop-lg-width) {
+    min-height: 567px;
+    padding: 5% 5% 4.5%;
+    border-radius: 50px;
+  }
+}
+
+.socials {
+  z-index: 2;
+  display: none;
+
+  @media (min-width: $desktop-md-width) {
+    position: absolute;
+    bottom: 60px;
+    right: 60px;
+    display: flex;
+    align-items: center;
+    content: '';
+  }
+}
+
+.socials:hover ~ .historyCard::before {
+  transform: scale(1.2);
+}
+
+.socialLink {
+  @include font(normal, 20px, 142%, 0.004em);
+
+  position: relative;
+  color: var(--third-text-color);
+  text-decoration: none;
+}
+
+.socialLink:not(:first-child) {
+  margin-left: 35px;
+}
+
+.socialLink:hover {
+  cursor: pointer;
+}
+
+.socialLink:hover:after {
+  position: absolute;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  height: 1px;
+  background-color: var(--third-text-color);
+  content: '';
+}
+
+.socialLink:active {
+  color: var(--active-link-color);
+}
+
+.socialLink:active:after {
+  background-color: var(--active-link-color);
+}
+
+.canvas {
+  position: absolute;
+  top: 0;
+  left: 0;
+  display: flex;
+  width: 100%;
+  min-height: 455px;
+  box-sizing: border-box;
+  border: none;
+  border-radius: 20px;
+  background-size: 100%;
+
+  @media (min-width: $desktop-md-width) {
+    top: 0;
+    left: 0;
+    min-height: 567px;
+    border-radius: 50px;
+  }
+
+  @media (min-width: $desktop-lg-width) {
+    top: 0;
+    left: 0;
+    min-height: 567px;
+    border-radius: 50px;
+  }
+}
+
+.historyCardTitle {
+  @include font(bold, 23px, 120%, 0.04em);
+  position: relative;
+  z-index: 2;
+  width: 60%;
+  margin: 2px 0 0 0;
+  pointer-events: none;
+
+  @media (min-width: $desktop-md-width) {
+    @include font(normal, 35px, 122%, 0.004em);
+
+    width: 100%;
+  }
+}
+
+.historyCardText {
+  @include font(normal, 16px, 142%);
+  position: relative;
+  z-index: 2;
+  width: 100%;
+  margin: 20px 0 0 0;
+  pointer-events: none;
+
+  @media (min-width: $desktop-md-width) {
+    @include font(normal, 20px, 142%, 0.0005em);
+
+    width: 70%;
+    margin-top: 35px;
+  }
+
+  @media (min-width: $desktop-lg-width) {
+    width: 55%;
+  }
+}
+
+.historyCardList {
+  @include font(normal, 16px, 142%);
+
+  position: relative;
+  margin: 25px 0 0 24px;
+  padding: 0;
+  list-style-type: none;
+  pointer-events: none;
+
+  @media (min-width: $desktop-md-width) {
+    @include font(normal, 20px, 142%);
+
+    margin: 26px 0 0 30px;
+  }
+}
+
+.historyCardItem {
+  position: relative;
+}
+
+.historyCardItem:before {
+  position: absolute;
+  top: 0;
+  left: -16px;
+  font-size: 30px;
+  color: var(--third-text-color);
+  content: '·';
+
+  @media (min-width: $desktop-md-width) {
+    left: -20px;
+    font-size: 38px;
+  }
+}
+
+.historyCardFooter {
+  position: relative;
+  display: flex;
+  justify-content: space-between;
+  margin-top: auto;
+  pointer-events: none;
+}
+
+.historyCardButton {
+  margin: 0 0 6px -15px;
+
+  @media (min-width: $desktop-md-width) {
+    margin: 0 0 0 -20px;
+  }
+}
+
+.controls {
+  display: flex;
+  justify-content: space-around;
+  margin: 20px;
+}
+
+.arrow {
+  width: 35px;
+  height: 35px;
+  fill: var(--main-text-color);
+}
+
+.arrow:hover {
+  fill: var(--btn-bg-color);
+}
+
+.button {
+  border: none;
+  background: none;
+  padding: 0;
+}
