+++ app/_components/SectionHistories/HistorySlider/index.js
@@ -1,0 +1,3 @@
+import HistorySlider from './HistorySlider';
+
+export default HistorySlider;
